Moving syntax object, fixing related issues
