+++ rholang/src/main/scala/coop/rchain/rholang/package.scala
@@ -1,7 +1,11 @@
 package coop.rchain
 
 import coop.rchain.metrics.Metrics
+import coop.rchain.rholang.interpreter.RhoRuntimeSyntax
 
 package object rholang {
   val RholangMetricsSource: Metrics.Source = Metrics.Source(Metrics.BaseSource, "rholang")
+  object syntax extends AllSyntaxRholang
+
 }
+trait AllSyntaxRholang extends RhoRuntimeSyntax
